Add changes required by CR
